split wallet.ts into files
