+++ src/transaction.ts
@@ -5,7 +5,6 @@ import {
   RecipientInterface,
 } from './types';
 import { Psbt, networks, address as laddress } from 'liquidjs-lib';
-import { estimateTxSize } from './wallet';
 
 export interface BuildTxArgs {
   psetBase64: string;
@@ -212,4 +211,78 @@ function toOutputScriptWithoutNetwork(address: string): Buffer {
       throw new Error('Invalid address');
     }
   }
+}
+
+// estimate segwit transaction size in bytes depending on number of inputs and outputs
+export function estimateTxSize(numInputs: number, numOutputs: number): number {
+  const base = calcTxSize(false, numInputs, numOutputs, false);
+  const total = calcTxSize(true, numInputs, numOutputs, true);
+  const weight = base * 3 + total;
+  const vsize = (weight + 3) / 4;
+
+  return vsize;
+}
+
+function calcTxSize(
+  withWitness: boolean,
+  numInputs: number,
+  numOutputs: number,
+  isConfidential: boolean
+) {
+  const inputsSize = calcInputsSize(withWitness, numInputs);
+  const outputsSize = calcOutputsSize(isConfidential, numOutputs);
+
+  return (
+    9 +
+    varIntSerializeSize(numOutputs) +
+    varIntSerializeSize(numInputs) +
+    inputsSize +
+    outputsSize
+  );
+}
+
+function calcInputsSize(withWitness: boolean, numInputs: number): number {
+  // prevout hash + prevout index
+  let size = (32 + 8) * numInputs;
+  if (withWitness) {
+    // scriptsig + pubkey
+    size += numInputs * (72 + 33);
+  }
+
+  return size;
+}
+
+function calcOutputsSize(isConfidential: boolean, numOutputs: number): number {
+  // asset + value + empty nonce
+  const baseOutputSize = 33 + 33 + 1;
+  let size = baseOutputSize * numOutputs;
+
+  if (isConfidential) {
+    // rangeproof + surjectionproof + 32 bytes for nonce
+    size += (4174 + 67 + 32) * numOutputs;
+  }
+
+  // fee asset + fee empty nonce + fee value
+  size += 33 + 1 + 9;
+
+  return size;
+}
+
+function varIntSerializeSize(val: number): number {
+  const maxUINT16 = 65535;
+  const maxUINT32 = 4294967295;
+
+  if (val < 0xfd) {
+    return 1;
+  }
+
+  if (val <= maxUINT16) {
+    return 3;
+  }
+
+  if (val <= maxUINT32) {
+    return 5;
+  }
+
+  return 9;
 }
