+++ test/transaction.test.ts
@@ -8,10 +8,10 @@ import {
 } from './fixtures/wallet.keys';
 import { APIURL, broadcastTx, faucet, mint, sleep } from './_regtest';
 import { buildTx, BuildTxArgs, decodePset } from '../src/transaction';
-import { fetchAndUnblindUtxos } from '../src/wallet';
 import * as assert from 'assert';
 import { RecipientInterface } from '../src/types';
 import { greedyCoinSelector } from '../src/coinselection/greedy';
+import { fetchAndUnblindUtxos } from '../src/explorer/esplora';
 
 jest.setTimeout(50000);
